chore: sepreated spec files for consistency
